requirements.txt
pandas==0.25.1
numpy==1.17.2
opencv_python==4.1.2.30
matplotlib==3.1.1
tqdm==4.36.1
six==1.12.0
ipython==7.13.0
Pillow==7.1.1
pydotplus==2.0.2
scikit_learn==0.22.2.post1
tf_object_detection_util==0.1.8
tinydb==3.15.2
